Split `GlobalTime.Model` into 2 files
