+++ JamGame/Assets/Scripts/Level/GlobalTime/TimeMeasures.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 50e2ee0a1fbb5e841a985b7d99c7731b
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
